+++ CHANGELOG.md
@@ -9,6 +9,8 @@ _Unreleased_
 
 - `rye version` can bump version by `--bump` option now.  #298
 
+- Fixed members not handled correctly in workspaces.  #300
+
 - Add `--clean` for `build` command.  #297
 
 - Fixed an issue where pip was not invoked from the right working directory